telegram_relay_bot/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==22.1.0
black==22.10.0
certifi==2022.9.24
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
discord.py==1.7.3
idna==3.4
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pyaes==1.6.1
pyasn1==0.4.8
pyperclip==1.8.2
requests==2.28.1
rsa==4.9
Telethon==1.25.4
typing_extensions==4.4.0
urllib3==1.26.12
yarl==1.8.1
